fix(dash): show embedded dashboard in Smapp
